Added Excel functionality
Stats for one session will now fill an Excel spreadsheet, filling the tables as necessary.

However, data ranges will not autofit, the series are renamed to default (Series1, Series2, ...), and macro-enabled workbooks (*.xlsm) will not raise an error with openpyxl, but will corrupt the file and make it unreadable to Excel.