Edge case handled
In the most recent session, the players played 5 hands total of PLO. Cedric only played one hand in one position, but played zero hands in the other position, causing a multitude of DivideByZero errors that I did not predict when the stat counts are divided by the number of hands in a given position.

I fixed this and avoid those errors by manually setting those percent values to -1 as a flag to avoid the division. The program also will not average those flag values, but will instead only use legitimate stats from positions actually played by the player.

This edge case occurs when players play very few hands of the desired poker type (or even combined), but will likely not occur very often.

Since the program currently only exports average stats to Excel (not positional), this edge case now makes sure output to Excel is correct. However, this is not so for prettyTable and the command line output, which shows a bunch of "-1". This will be handled in the next commit.