Date convention changed
Date convention is now more conventional, in the format "MMDD". Year will be added when a session is played past April 9, 2022. However, I should add this functionality now while it is fresh, so that I won't run into any issues if I am still using this April of next year.