GUI fully functional
GUI is now fully functional. It can take input, print output to the user. There are no messages for crashes or errors, but those will show up in the command prompt anyway.

Command line prompts were removed because all information is now output to the GUI. Any crashes will appear in the command prompt still, however.

After a few aesthetic changes and a loading/task complete message, the GUI will be complete and the program ready to use.