Changed WASD
WASD is now shown both as an absolute statistic (divisor is the total hands played) and as a relative statistic to times went to showdown (divisor is WTSD count).

This new change is reflected in the current sessions stats outputs (Excel and command line), and will be reflected in stats over time once I refill the chart.

Old logs and ledgers were deleted since new ones work just fine with the program.