Logs can either run as-is or reversed
PokerNow exports logs in ascending order (latest is at the top). The program can now handle this order, instead of having to reverse the order every time. However, a hard-coded exceptions list is available to read logs in descending order. This is for debugging purposes, as it is much easier to read the logs in descending order.