Functions moved to fewer files
My last commit from the Jackson Square office. It has been a good run. I've enjoyed writing this program from scratch. I hope it runs bug free and is useful forever, to me and to others.

There are many helper functions in the PokerStats program. Before now, each had their own file, as in MATLAB. Now, a bunch of functions are in a few files, for conciseness.